hotfix: tensor data types, bug in inference code

+++ datahub/datahub/transforms/template.py
@@ -823,7 +823,7 @@ def featurize_templates_like_af3(
         boundaries=torch.as_tensor(distogram_bins, dtype=template_distogram.dtype, device=template_distogram.device),
     )
     n_bins = len(distogram_bins) + 1
-    template_distogram = torch.nn.functional.one_hot(template_distogram, num_classes=n_bins)
+    template_distogram = torch.nn.functional.one_hot(template_distogram, num_classes=n_bins).to(torch.float32) # We don't need int64 precision
 
     return {
         "template_restype": res_type,
